requirements.txt
asf-search
python-dotenv
Shapely
rasterio
scikit-image
numpy
scipy
matplotlib